Update transaction pagkace
* Split into files by query
* Improve error
